Move dmatest.h dependency to fix build issues
